restructure index folder
