+++ lego-webapp/pages/index/AuthenticatedFrontpage.tsx
@@ -8,6 +8,12 @@ import Banner from '~/components/Banner';
 import HsSectionContent from '~/components/HsSection/HsSection';
 import Poll from '~/components/Poll';
 import RandomQuote from '~/components/RandomQuote';
+import CompactEvents from '~/pages/index/_components/CompactEvents';
+import ArticleItem from '~/pages/index/_components/authenticated/ArticleItem';
+import FrontpageEventItem from '~/pages/index/_components/authenticated/FrontpageEventItem';
+import LatestReadme from '~/pages/index/_components/authenticated/LatestReadme';
+import Pinned from '~/pages/index/_components/authenticated/Pinned';
+import UpcomingRegistrations from '~/pages/index/_components/authenticated/UpcomingRegistrations';
 import { fetchCurrentPrivateBanner } from '~/redux/actions/BannerActions';
 import { fetchData, fetchReadmes } from '~/redux/actions/FrontpageActions';
 import { fetchRandomQuote } from '~/redux/actions/QuoteActions';
@@ -25,13 +31,7 @@ import { selectPinnedPoll } from '~/redux/slices/polls';
 import { selectRandomQuote } from '~/redux/slices/quotes';
 import utilStyles from '~/styles/utilities.module.css';
 import { guardLogin } from '~/utils/replaceUnlessLoggedIn';
-import ArticleItem from './ArticleItem';
 import styles from './AuthenticatedFrontpage.module.css';
-import CompactEvents from './CompactEvents';
-import FrontpageEventItem from './FrontpageEventItem';
-import LatestReadme from './LatestReadme';
-import Pinned from './Pinned';
-import UpcomingRegistrations from './UpcomingRegistrations';
 import { itemUrl, renderMeta } from './utils';
 import type { EntityId } from '@reduxjs/toolkit';
 import type { FrontpageEvent } from '~/redux/models/Event';
