+++ lego-webapp/pages/index/_components/CompactEvents.module.css
