+++ lego-webapp/pages/index/_components/authenticated/FrontpageEventItem.tsx
@@ -3,7 +3,7 @@ import { colorForEventType } from '~/pages/events/utils';
 import { useAppSelector } from '~/redux/hooks';
 import { useIsLoggedIn } from '~/redux/slices/auth';
 import { eventStatus } from '~/utils/eventStatus';
-import styles from './EventItem.module.css';
+import styles from './FrontpageEventItem.module.css';
 import type { FrontpageEvent } from '~/redux/models/Event';
 
 type Props = {
