+++ lego-webapp/pages/index/_components/authenticated/UpcomingRegistrations.module.css
