+++ lego-webapp/pages/index/_components/public/Hero.tsx
