+++ lego-webapp/pages/index/_components/public/MainSponsor.tsx
