+++ lego-webapp/pages/index/_components/public/useTitleBeadAnimation.ts
